Add more deprecated methods, move all deprecations to separate file
