+++ lib/rocksdb.rb
@@ -26,7 +26,6 @@ module RocksDB
 
   class DB
     extend Forwardable
-    extend Gem::Deprecate
 
     def initialize(path, rocksdb_options = "", options = {})
       is_readonly = options[:readonly] || false
@@ -67,20 +66,7 @@ module RocksDB
 
     alias_method :each_index, :each_key
     alias_method :each_with_index, :each_pair
-
-    # Deprecated methods, kept for backward compatibility
-    alias_method :new_iterator, :to_iterator
-    deprecate :new_iterator, :to_iterator, 2019, 12
-
-    def is_readonly?
-      !writable?
-    end
-    deprecate :is_readonly?, :writable?, 2019, 12
-
-    alias_method :is_open?, :open?
-    deprecate :is_open?, :open?, 2019, 12
-
-    alias_method :multi_get, :get_many
-    deprecate :multi_get, :get_many, 2019, 12
   end
 end
+
+require 'rocksdb/ruby/deprecated'
